+++ src/infrastructure/RepositoryFactory.ts
@@ -20,7 +20,6 @@ import { NetworkType } from '../model/network/NetworkType';
 import { AccountRepository } from './AccountRepository';
 import { BlockRepository } from './BlockRepository';
 import { ChainRepository } from './ChainRepository';
-import { HashLockRepository } from './HashLockRepository';
 import { IListener } from './IListener';
 import { MetadataRepository } from './MetadataRepository';
 import { MosaicRepository } from './MosaicRepository';
@@ -31,9 +30,11 @@ import { NodeRepository } from './NodeRepository';
 import { ReceiptRepository } from './ReceiptRepository';
 import { RestrictionAccountRepository } from './RestrictionAccountRepository';
 import { RestrictionMosaicRepository } from './RestrictionMosaicRepository';
-import { SecretLockRepository } from './SecretLockRepository';
 import { TransactionRepository } from './TransactionRepository';
 import { TransactionStatusRepository } from './TransactionStatusRepository';
+import { HashLockRepository } from './HashLockRepository';
+import { SecretLockRepository } from './SecretLockRepository';
+import { FinalizationRepository } from './FinalizationRepository';
 
 /**
  * A repository factory allows clients to create repositories to access NEM Server without knowing
@@ -132,6 +133,11 @@ export interface RepositoryFactory {
      * @returns a newly created {@link SecretLockRepository}
      */
     createSecretLockRepository(): SecretLockRepository;
+
+    /**
+     * @returns a newly created {@link FinalizationRepository}
+     */
+    createFinalizationRepository(): FinalizationRepository;
 
     /**
      * @returns a newly created {@link IListener}