+++ src/infrastructure/RepositoryFactoryHttp.ts
@@ -16,7 +16,6 @@
 
 import { defer, Observable, of as observableOf } from 'rxjs';
 import { map, shareReplay } from 'rxjs/operators';
-import { DtoMapping } from '../core/utils/DtoMapping';
 import { NetworkCurrencies } from '../model/mosaic';
 import { NetworkConfiguration } from '../model/network/NetworkConfiguration';
 import { NetworkType } from '../model/network/NetworkType';
@@ -27,8 +26,6 @@ import { BlockHttp } from './BlockHttp';
 import { BlockRepository } from './BlockRepository';
 import { ChainHttp } from './ChainHttp';
 import { ChainRepository } from './ChainRepository';
-import { HashLockHttp } from './HashLockHttp';
-import { HashLockRepository } from './HashLockRepository';
 import { IListener } from './IListener';
 import { Listener } from './Listener';
 import { MetadataHttp } from './MetadataHttp';
@@ -51,13 +48,17 @@ import { RestrictionAccountHttp } from './RestrictionAccountHttp';
 import { RestrictionAccountRepository } from './RestrictionAccountRepository';
 import { RestrictionMosaicHttp } from './RestrictionMosaicHttp';
 import { RestrictionMosaicRepository } from './RestrictionMosaicRepository';
-import { SecretLockHttp } from './SecretLockHttp';
-import { SecretLockRepository } from './SecretLockRepository';
 import { TransactionHttp } from './TransactionHttp';
 import { TransactionRepository } from './TransactionRepository';
 import { TransactionStatusHttp } from './TransactionStatusHttp';
 import { TransactionStatusRepository } from './TransactionStatusRepository';
-
+import { HashLockRepository } from './HashLockRepository';
+import { SecretLockRepository } from './SecretLockRepository';
+import { SecretLockHttp } from './SecretLockHttp';
+import { HashLockHttp } from './HashLockHttp';
+import { DtoMapping } from '../core/utils/DtoMapping';
+import { FinalizationHttp } from './FinalizationHttp';
+import { FinalizationRepository } from './FinalizationRepository';
 /**
  * Receipt http repository.
  *
@@ -172,6 +173,10 @@ export class RepositoryFactoryHttp implements RepositoryFactory {
 
     createSecretLockRepository(): SecretLockRepository {
         return new SecretLockHttp(this.url, this.fetchApi);
+    }
+
+    createFinalizationRepository(): FinalizationRepository {
+        return new FinalizationHttp(this.url, this.fetchApi);
     }
 
     getGenerationHash(): Observable<string> {