+++ src/infrastructure/paginationStreamer/PaginationStreamer.ts
@@ -18,7 +18,7 @@ import { concat } from 'rxjs';
 import { Observable } from 'rxjs/internal/Observable';
 import { defer } from 'rxjs/internal/observable/defer';
 import { from } from 'rxjs/internal/observable/from';
-import { flatMap } from 'rxjs/operators';
+import { mergeMap } from 'rxjs/operators';
 import { SearchCriteria } from '../searchCriteria/SearchCriteria';
 import { Searcher } from './Searcher';
 
@@ -55,7 +55,7 @@ export class PaginationStreamer<E, C extends SearchCriteria> {
         return defer(() => {
             const observable = this.searcher.search(criteria);
             return observable.pipe(
-                flatMap((page) => {
+                mergeMap((page) => {
                     if (page.isLastPage) {
                         return from(page.data);
                     } else {