+++ src/model/finalization/Stage.ts
@@ -1,7 +1,7 @@
 /*
  * Copyright 2020 NEM
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
+ * Licensed under the Apache License, Version 2.0 (the "License"),
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
@@ -14,11 +14,12 @@
  * limitations under the License.
  */
 
-import { Searcher } from './paginationStreamer/Searcher';
-import { SecretLockInfo } from '../model/lock/SecretLockInfo';
-import { SecretLockSearchCriteria } from './searchCriteria/SecretLockSearchCriteria';
-
 /**
- * Secretlock interface repository.
+ * Message group stage enum
  */
-export type SecretLockRepository = Searcher<SecretLockInfo, SecretLockSearchCriteria>;
+
+export enum Stage {
+    Prevote = 0,
+    Precommit = 1,
+    Count = 2,
+}