+++ test/infrastructure/RepositoryFactory.spec.ts
@@ -16,12 +16,10 @@
 import { expect } from 'chai';
 import { of, of as observableOf } from 'rxjs';
 import { catchError, map } from 'rxjs/operators';
-import { NetworkConfigurationDTO } from 'symbol-openapi-typescript-fetch-client';
 import { instance, mock, when } from 'ts-mockito';
 import { AccountHttp } from '../../src/infrastructure/AccountHttp';
 import { BlockHttp } from '../../src/infrastructure/BlockHttp';
 import { ChainHttp } from '../../src/infrastructure/ChainHttp';
-import { HashLockHttp } from '../../src/infrastructure/HashLockHttp';
 import { Listener } from '../../src/infrastructure/Listener';
 import { MetadataHttp } from '../../src/infrastructure/MetadataHttp';
 import { MosaicHttp } from '../../src/infrastructure/MosaicHttp';
@@ -36,13 +34,16 @@ import { ReceiptHttp } from '../../src/infrastructure/ReceiptHttp';
 import { RepositoryFactoryHttp } from '../../src/infrastructure/RepositoryFactoryHttp';
 import { RestrictionAccountHttp } from '../../src/infrastructure/RestrictionAccountHttp';
 import { RestrictionMosaicHttp } from '../../src/infrastructure/RestrictionMosaicHttp';
-import { SecretLockHttp } from '../../src/infrastructure/SecretLockHttp';
 import { TransactionGroup } from '../../src/infrastructure/TransactionGroup';
 import { TransactionHttp } from '../../src/infrastructure/TransactionHttp';
 import { TransactionStatusHttp } from '../../src/infrastructure/TransactionStatusHttp';
 import { NetworkCurrencies } from '../../src/model/mosaic/NetworkCurrencies';
 import { NetworkType } from '../../src/model/network/NetworkType';
 import { NodeInfo } from '../../src/model/node/NodeInfo';
+import { HashLockHttp } from '../../src/infrastructure/HashLockHttp';
+import { SecretLockHttp } from '../../src/infrastructure/SecretLockHttp';
+import { FinalizationHttp } from '../../src/infrastructure/FinalizationHttp';
+import { NetworkConfigurationDTO } from 'symbol-openapi-typescript-fetch-client';
 
 describe('RepositoryFactory', () => {
     it('Should create repositories', () => {
@@ -66,6 +67,7 @@ describe('RepositoryFactory', () => {
         expect(repositoryFactory.createTransactionRepository()).to.be.not.null;
         expect(repositoryFactory.createHashLockRepository()).to.be.not.null;
         expect(repositoryFactory.createSecretLockRepository()).to.be.not.null;
+        expect(repositoryFactory.createFinalizationRepository()).to.be.not.null;
     });
 
     it('Should get GenerationHash from cache', (done) => {
@@ -329,6 +331,7 @@ describe('RepositoryFactory', () => {
         expect(factory.createTransactionStatusRepository() instanceof TransactionStatusHttp).to.be.true;
         expect(factory.createHashLockRepository() instanceof HashLockHttp).to.be.true;
         expect(factory.createSecretLockRepository() instanceof SecretLockHttp).to.be.true;
+        expect(factory.createFinalizationRepository() instanceof FinalizationHttp).to.be.true;
     });
 
     it('Fail remote call ', async () => {