+++ CHANGELOG.md
@@ -11,11 +11,15 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 ---|---|---
 SDK Core| v0.21.1 | [symbol-sdk](https://www.npmjs.com/package/symbol-sdk)
 Catbuffer | v0.0.22 | [catbuffer-typescript](https://www.npmjs.com/package/catbuffer-typescript)
-Client Library | v0.10.0-3  | [symbol-openapi-typescript-fetch-client](https://www.npmjs.com/package/symbol-openapi-typescript-fetch-client)
+Client Library | v0.10.1  | [symbol-openapi-typescript-fetch-client](https://www.npmjs.com/package/symbol-openapi-typescript-fetch-client)
 
 - **[BREAKING CHANGE]** `Deadline.create` requires the configurable `epochAdjustment` from the network properties.
 - **[BREAKING CHANGE]** `NetworkCurrency` subclasses replaced with `Currency` objects.
+- **[BREAKING CHANGE]** `SecreatLockRepository.getSecretLock` has been removed. You can now search by secret by using the search criteria.
+- Added `FinalizationRepository`.
+- Added `transferMosaicId`, `fromTransferAmount`, `toTransferAmount` to transaction searches.
 - Added `CurrencyService` to allow loading Network and custom `Currency` objects from the rest service.
+- Added `UnlockedAccount` endpoint in `NodeRepository` for checking delegated harvesting status.
 
 ## [0.21.0] - 25-Sep-2020